Put the three steps of the pass on the command line
The workflow is three commands with a person between the first two. Grouping probe
records the setup under every configuration of its probe list and writes the workbook;
grouping import normalises the filled-in workbook into the file that is committed; and
record --grouping builds the grouped energy system from it and reports both what it
proved and what it could not. All three take the setup and work the rest of the paths
out, because a probe list, a workbook, a decision and a grouped file are all named for
the setup and all live beside the recorded twin.

The verbs live in their own module. The command line had already grown past six hundred
lines and these are a third as much again, so what moves out with them is the two
renderers — most of the old file, and about the layout of a page of output rather than
about the shape of the command line — and the exit codes, which everything needs and
which would otherwise make the renderers and the verbs import each other. What stays is
the parser, the dispatch and the verbs themselves, so the shape of the command line is
still readable in one place. Nothing changes about what any existing command does.
