+++ mypy_moderate.ini
@@ -108,3 +108,6 @@ ignore_missing_imports = True
 
 [mypy-jsonschema.*]
 ignore_missing_imports = True
+
+[mypy-openpyxl.*]
+ignore_missing_imports = True
